CORE-5960 Bad CPI upload returns 400 http status (#1781)
Return 400 for bad uploads that have triggered a `ValidationException`

Otherwise return 500 for an internal server error.

Moved `ValidationException` so it can be accessed from both the
chunk-db-writer-impl, and rpc ops layer.  Then we check on the FQN
name of the exception, if thrown, to return the appropriate http
status code.