CORE-5964 - Switch p2p to use Corda's domain model HoldingIdentity (#1783)
Currently, p2p code was predominantly using the Avro-type `net.corda.data.identity.HoldingIdentity` in internal logic. This change switches the code to use the `net.corda.virtualnode.HoldingIdentity` type, instead. The latter models the x500 name as a strong type (`MemberX500Name`), instead of a string. This ensures any comparisons will be order-insensitive and reliable and any hashing will be stable.

This was introducing a bug in the p2p side, where if a message was sent to an identity where the x500 name had a slightly difference to the one registered in the members map (e.g. the attributes in the string were in different order), then p2p would classify those as different identities and would not send the message.