CORE-6035 Ensure MGM session key hash is available to members (#1799)
Members receiving messages from the MGM needs to lookup the member info by session key hash. Currently we are not setting that on the mgm info loaded from the group policy file so lookup fails and message sending fails. Added session key hash set up to MGM onboarding. Also added back up hash calculation if hash is not set. This will log a warning also so that we can see when it occurs. It is better to store the hash rather than calculate so that lookups can be performed faster.