CORE-6044: allocate port dynamically for health monitor tests (#1808)
Previously we just listened on port 7000, hoping it was free. On my work MacOs Monterey
machine Control centre, which is a standard part of MacOS, is already using port 7000
so the tests fail. Even if that were not the case, assuming one specific TCP port
is available introduces a race condition if s multiple test cases are running in parallel.
So, we let the OS pick the port, for test purposes.