CORE-6050 - Reuse connection pool in membership components (#1803)
Change usage from `createEntityManagerFactory` to `getOrCreateEntityManagerFactory` to force the usage of the connection pool cache (`DbConnectionOpsCachedImpl`).

The issue was that every time we accessed the database we created a new Hikary pool with ten new connections and didn't close it. This cause PostgreSQL to use all the connections after a few calls. The exception is ` ERROR com.zaxxer.hikari.pool.HikariPool {} - HikariPool-79 - Exception during pool initialization.
org.postgresql.util.PSQLException: FATAL: remaining connection slots are reserved for non-replication superuser connections`

Testing the change shows that it removes the exception and allows reusing the pool multiple times.