CORE-5455  Strongly typed short hash for holding id. (#1802)
Use a strongly typed short hash to reference the holding identity and pass back and forth to the client, and within the code.

Using a regular class and not a data class because we want the class to only contain valid constructed objects so we are:

    - using a companion object method of(...) to validate the the string we are passed
    - use a private constructor to create the ShortHash if it is correct
    - data class and the copy() method expose private constructor (https://youtrack.jetbrains.com/issue/KT-11914/Confusing-data-class-copy-with-private-constructor), so we implement equals etc. (basically it's a "synthetic" data class).

ShortHash is created via the companion object method of. In non-test code there should be very few calls to ShortHash.of(...). We expect to generate a ShortHash from incoming client REST calls, and that's it.

Using ShortHash.value (the field that contains... the short hash String) in most places where required. Generally, comparing value anywhere is an 'anti-pattern' and needs fixing (probably outside the scope of this PR).

The value is upper case to match what we return from SecureHash.toHexString() (even though we can construct from a lowercase hex string).

toString() returns .value as there a few places where we do things like

val s = "something_$shortHash"

Several cases of misuse of the holding identity short hash when accessing virtual node db connections.

The ShortHash throws its own ShortHashException if it cannot be created from a 12+ character hexstring.