CORE-5992 FlowContext implementation (#1843)
Created a Corda internal interface for flattening properties to encode into SessionInit and FlowStartContext messages

Implemented FlowContextImpl, a stackable context properties implementation

FlowStack api added to allow stack items to be pushed with context

FlowContextImpl is entirely backed by containers in items in the FlowStack, which means they will automatically be serialized and deserialized as Flows are suspended and resumed. It also means as sub flows are added to, and removed from the stack, their context is scoped accordingly.