CORE-6111 - Adjust link manager to treat cases where destination is not in the network yet gracefully (#1847)
Introducing an extra conditional for cases where the destination identity is not part of the network yet. In case of unauthenticated messages, the message is just dropped with a warning log. In case of authenticated message, the message is kept for replaying. Previously, we were not processing replays for scenarios where the destination identity is locally hosted. This logic is removed now. That should be safe, since the issue we were trying to mitigate is the link manager depositing the message more than once, but that should not be an issue as p2p does not provide exactly-once delivery guarantees. The possibility of that happening has also been reduced signficantly now that we have two type of markers (processed / sent) and the first one (that replays are based on) is only written only once atomically with the consumption of events from `p2p.out`.