CORE-6142 - bugfix allow multiple duplex channels in websocket adaptor (#1908)
Asserted this fixes the issue mentioned in the jira, connect 10 websocket connections for a flow, wait for them each to time out, notice that the correct session is disconnected and correct listeners removed. After they disconnect we can connect again to the flow (prior to the fix, we got error "max 10 connections...").