CORE-4424 Port Link Manager to common worker framework (#1902)
Created a P2P Link Manager worker which uses the common worker framework and delete the legacy Link Manager application. The docker image has been renamed from `corda-os-p2p-link-manager` to `corda-os-p2p-link-manager-worker`. I have updated the P2P deployment to use the new docker image and to pass in the address of the Kafka Broker as a command line argument instead of an environment variable. Using an environment variable was incompatible
with the common worker framework.