NOTICK Return early in `CordaRPCSenderImpl` on serialization error (#1913)
`CordaRPCSenderImpl.sendRequest` was causing a null pointer error if the
request failed to serialize. This was due to the method continuing on
even though the future had been completed exceptionally. The error was
masking the underlying error causing the serialization failure.

Returning early when there is a serialization error resolves this.