CORE-6081 Bulk persistence operations (#2015)
Implement bulk operations in `PersistenceServiceImpl`.

Due to `FindEntity` changing to `FindEntities`, the code in `EntityMessageProcessor` and `PersistenceServiceInternal` needed to change to accommodate multiple entities being passed in.

All methods in `PersistenceServiceImpl` now use `requireBoxedType` to ensure that no primitives are passed in, as they either fail on the db-worker side anyway or cause serialization/casting errors in the flow-worker. Furthermore, primitives cannot be entities, so the check fits even based on that assertion.