CORE-6651 Remove `FlowException` (#2055)
`FlowException` is not being used anymore, since we are not encouraging
the sending of exceptions between sessions to communicate errors.
Instead they should pass around POJOs that signify an error.

`FlowException`s specified an exception as needing to be seen as it was
on a peer flow. This behaviour is no longer required.