CORE-4869 - Reject messages where source and destination group IDs do not match (#2091)
## Changes
This change introduces logic that rejects messages where source and destination group IDs do not match. This will be adjusted in the future, if/when we get to implement cross-group communication. I also adjusted the `P2PLayerEndToEndTest`, so that it's easy to temporarily adjust it to test communication across groups.

This makes use of a new marker introduced in: https://github.com/corda/corda-api/pull/569

## Testing
Apart from the added unit tests, I temporarily adjusted the `P2PLayerEndToEndTest` to run the test with the locally hosted identities across groups. Before the change, it was passing. After the change, it started failing as expected. I reverted the change and didn't introduce a new test there as I think capturing that only at the unit level is reasonable.