CORE-6698 Vnode reconciler goes down (#2092)
The problem was, Vnode reconciler was going down.

Vnode reconciler component depends on `VirtualNodeInfoWriterComponentImpl` (`VirtualNodeInfoWriteService` is used as the Kafka writer in Vnode info reconciliation).

In `VirtualNodeInfoWriterComponentImpl.onConfigChangedEventReceived` we would switch coordinator status to DOWN every time we would receive a `ConfigChangedEvent` to mark that component is down during resources re-creation. However, that is propagated to components depending on this one and therefore to Vnode reconciler.

* removes `VirtualNodeInfoWriterComponentImpl` switching to DOWN every time we receive a `ConfigChangedEvent`.

* improves logging for Vnode reconciler readers and writer.