CORE-6762: Ensure calling getDefaults does not break default insertion for subsequent calls to the validator (#2141)
There appears to be some sort of issue with the config validation library used by C5, whereby if the first walk of a schema is performed with validation turned off, subsequent walks may result in NPEs when trying to default objects. This was causing intermittent failures in the build. The getDefaults function does this, and so the validation library could break if a call to that function was the first usage of it. As this is used to get the default config for connecting to the database, an early call here is somewhat likely.

The fix here is to turn on validation even when running getDefaults, which avoids the problem. (This makes getDefaults equivalent to validate with applyDefaults enabled against an empty config.)