CORE-5266 Exception Logging Cleanup (#2174)
Exceptions are propagated to flows which didn't generate them in order that they can know to discontinue, and they have to discontinue by throwing an exception. Originally these exceptions were re-thrown as-is by all flows as any other they would generate internally, which meant callstacks were logged not only for the flow where the error occurred, but the ones which were discontinuing, and those "discontinuing" callstacks always pointed to internal error handling code. This was misleading for user code, and could obfuscate the root cause of any problems: it made it look like Corda was broken rather than user code.

Instead we introduce an explicit discontinuation exception which wraps the root cause. This can be caught as a special case and the user is only informed that the flow discontinued under Corda's defined error handling circumstances.