CORE-6619: move composite key implementation out of API
This is a step toward removing Bouncy Castle from our public
API in corda-api, and also gets some of the implementation detail
(and quite a few lines of code) out of the public API.

Make the logic that decodes ASN1 content in to a composite key
be private, so that it can only be used via the standard
Java Security key signing approach. This involved deleting a
couple of tests that covered the internal logic directly, since
it is no longer accessible. We already had the exact same
coverage in system tests which involve Java Security.