CORE-5088: (Start of) consensual ledger persistence (#1883)
CORE-5088: Consensual ledger persistence (initial version)

Added "ledger-persistence-service" and "ledger-persistence-service-impl" modules that services custom persistence requests for the consensual ledger in the DB worker. A new message processor ("ConsensualLedgerMessageProcessor") receives requests for persistence operations on the consensual ledger. The requests contain a payload of serialised transaction data. A database repository ("ConsensualLedgerRepository") is called by the message processor to read or write to the database with native JPA queries. Implemented queries to retrieve transaction by ID and to persist transaction.

Some code from "entity-processor-service-impl" was reused so this module was refactored and common code was moved into modules "persistence-service-common" and "persistence-testkit".