NOTICK: Update test dependencies (#2254)
The following dependencies were updated:
- assertJ 3.12.2 -> 3.23.1
- hamcrest 2.1 -> 2.2
- junit5 5.9.0 -> 5.9.1
- junit-platform 1.9.0 -> 1.9.1
- mockito 4.6.1 -> 4.8.0
- osgiJunit5Test 1.1.0 -> 1.2.1
- sling 3.3.0 -> 3.3.2

I had spotted a workaround for the issue mentioned in the comment in the gradle file for assertJ, which is why I've attempted to migrate here. It turns out there is more to it though, as assertJ regularly deprecates APIs and this causes compilation errors for us (due to warnings-as-errors being enabled). Everything else in the diff is either this workaround or removal of deprecated APIs.