CORE-6316 - Make netty buffer allocation dynamic (#2227)
Adjusting the gateway, so that the server dynamically expands the buffer to hold the request payload, instead of allocating everything eagerly. Note: this required moving off from the `buffer.readBytes()` API of Netty to the `buffer.writeBytes()`.

While working at it & testing, I realised that in some cases where requests are invalid, we were keeping the TPC connections open. So, I slightly adjusted this to close the connections eagerly in these scenarios.