CORE-7091 Add service for retrieving the platform version (#2296)
Added a new internal service which exposes the active platform version of the current corda cluster, the local worker platform version, and the local worker software version. This service is in a new libs module called platform-info. Currently the platform version values are hard coded as a stepping stone to unblock teams which need an API to work against. The software version is available through the manifest file so the service returns the actual version and not a stub.

This also updates the membership registration services, which previously hard coded the active platform version, to use the new service.

Other changes are tidy up items such as updating tests to use a more realistic value for platform version, and remove unused properties.

As discussed internally, the implementation for platform version is just a temporary solution for now.