Rework to attach a changeID as an lb tag
This commit also contains some early from of e2e test, as well as
migrating some of the code around changelog fetching to use DTOs as it
allows injecting a change ID into the actual entities, without making
the property of the entity a variable. Additional might be needed for
later work against changelogs.