CORE-6149: Add integration and unit tests for notary JPA backing store (#2211)
CORE-6149: Add integration and unit-tests for Notary JPA backing store. 

Note that emphasis is made on integration tests as we want to test against a real DB. There are currently two disabled tests either because the expected behaviour is not clear or the test fails. Both need to be reviewed in the ticket provided in the comment. There are two other tests that pass reflecting the current logic but also need review in a separate task.