CORE-6083 - Add URL path as a config option for gateway (#661)
Add configuration option to adjust the URL path the gateway server is listening to for requests.

To keep things simple, I've added a validation pattern that allows relatively simple paths for now (i.e. alphanumerics separated by slashes). The root path was selected as default, so that it works out of the box with URLs that do not specify any custom path.