CORE-7027 Transaction verification in consensual finality flows (#2411)
Catch `IllegalStateException`s, `IllegalArgumentException`s and
`CordaRuntimeException`s thrown from the
`ConsensualSignedTransactionVerifier` passed into
`ConsensualReceiveFinalityFlow`.

If one of these exceptions is thrown, then the failure is communicated
back to the other side of the session (`ConsensualFinalityFlow`) causing
it to end. The whole process of finality then ends as the
`ConsensualFinalityFlow` will propagate errors to all sessions involved
in finality.

Add `Payload` is a way to communicate between flows if successful
processing or an error occurred.