CORE-7284 Set version information in the MemberInfo (#2391)
Ensures that all MemberInfo's generated by dynamic member registration, static member registration and mgm registration report their version information. This includes:
* Platform version
* Software version
* CPI Name
* CPI Version
* CPI Signer Summary Hash

The changes in the test class can be difficult to see in the diff because I moved the unit tests to the nested class structure to make it easier to group and find tests. All tests are identical other than:
* DynamicMemberRegistrationServiceTest
> "registration context is constructed as expected": New test
> "registration fails if virtual node info cannot be found": New test
* MGMRegistrationServiceTest
> "registration successfully builds MGM info and publishes it": Added extra assertions for new properties
> "registration fails when vnode info cannot be found": New test
* StaticMemberRegistrationServiceTest
> "during registration, the registering static member inside the GroupPolicy file gets parsed and published": Added extra assertions for new properties
> "registration fails when virtual node info is unavailable": New test

The MGM registration became quite complex so I've refactored that in to smaller classes and added tests for each.