CORE-6175  Handle InvalidProducerEpochException in the CordaKafkaProducer (#2418)
Kafka can give us an InvalidProducerEpochException if the broker increments the producer epoch. Ordinarily this would happen if a new producer arrives with the same transactional.id (which for us should only happen across a replica restart or component restart). That case would result in a ProducerFencedException if the old producer is still around and an attempt is made to use it. However, this can also happen if the broker times the transaction out, which can happen if the client is temporarily unresponsive (e.g. via long GC cycles). In this case, an InvalidProducerEpochException is produced instead. In this case, the current transaction should be aborted, the underlying producer recreated, and the transaction retried.