CORE-4442 Add support for session certificates (#642)
Add the session certificate to the encrypted extensions of both Initiator Handshake and Responder Handshake. Add a revocation option to the LinkManager config. So the revocation check (of the session certifcates) can be controlled in the Link Manager in the same way as the Gateway (for TLS certificates). Add the session certificates to the locally hosted identity.