CORE-4442 Add support for session certificates (#2286)
Change the authentication protocol to check that session certificates are valid. Send the session certificates across the wire as apart of Diffie-Hellman key exchange in the (initiator/responder) handshake messages. Verify the certificates are valid when we receive said messages. To do this I added the certificates to the locally hosted identities. I changed the RPC command which publishes this data to Kafka so they can be included.