CORE-6885 Filtered transactions (#2408)
Add `FilteredTransaction` and its implementation, along with `FilteredTransactionFactory`, which creates an instance from a `WireTransaction`.

`FilteredTransactionFactory.create` takes a `WireTransaction` and a list of `ComponentGroupFilterParameters` which specify which component groups to keep and what type of proof should be used if kept. The existence of a `ComponentGroupFilterParameters` for a specific index means that the filtered transaction will include the component group with that index from the `WireTransaction`.

`FilteredTransactionImpl` contains a lot of verification code around the filtered components that exist within it.

Currently, `FilteredTransaction` remains as an internal API, but that may change in the future.