CORE-5561 - Handle scenarios where source identity in message is not locally hosted (#2450)
Changes
When the source identity of an authenticated or unauthenticated message is not locally hosted, messages are dropped.

Testing
In addition to unit tests added this was tested with a manual deployment. The sending app simulator sent 24 messages of size 10000 bytes, in batches of 1, at 10 second intervals. All of these messages were received.