CORE-7843 No such requestId reported whilst executing e2e tests (#2487)
Caught exception in order to retry correctly.

The uploadCpi function attempts to retry failed requests using an "eventually" loop. However one omission was that it relies on the status not being OK in order to "fail" and then spin the loop again. The problem was the remote client status() method it uses actually throws an exception in the circumstances the status Http request fails - when there is a not found request Id. Throwing exited the loop instead of letting it spin again, so that has to be caught in this one expected case.