CORE-7844 Increase stability of the config update endpoint (#2500)
When a config request is sent with the section set to "corda.messaging" nearly all components in the system will respond to this config change. This will cause the HttpGateway to go down bringing down the HttpServer. This will close all the connections open from http clients. this means a response is not received by the client for the request. To avoid this in tests we swallow the UniRestException thrown for put /config and does not retry the PUT, Previously the test would retry which caused multiple messaging config events to be sent to the RPC worker.

Separately the RPC pattern when sending config to the db worker will also fail due to repartition exceptions thrown by the RPC messaging pattern as a result of the messaging config being updated. A tmp solution is added to return the config posted by the client aswell as a success response.

A better solution might be to not bother waiting for a response from the db worker for config updates and immediately return 202 Accepted to the client always