CORE-6690 consistent overrides (#2364)
Introduce consistency in the way in which credentials are handled in the overrides. Standardize on `username` over `user`. For each field that represents part of a credential, allow it to be specified directly or via secret. For example:

```
db:
  cluster:
    username:
      value: "user"
```

or

```
db:
  cluster:
    username:
      valueFrom:
        secretKeyRef:
          name: "my-db-secret"
          key: "username"
```

Also, move the Postgres client image configuration from `db.clientImage` to `bootstrap.db.clientImage` to reflect the fact that it is only used at bootstrap time.

Co-authored-by: David Currie <david.currie@r3.com>