CORE-7999 force flow tests to run last so they do not interfere with other tests.  (#2522)
The flow test cluster configuration changes are picked up and workers continue to operate normally updated the messaging config as a final step. This triggers a restart of many components such as RPC message pattern responders. Subsequent tests such the ConfigTests are reliant on these being UP and so they can fail due to the race condition. Ordering this test to be last to reduce impact.