CORE-6962 adjust component logging and lifecycle coordinator logging (#2521)
This PR tries to standardize how components log their lifecycle statuses as well as how they log tracked registrations. Logging is centralized in the LifecycleProcessor and Registration classes. Surplus logging is removed from individual components as a result. Numerous other logs are moved to debug or trace which are not essential.