CORE-8041: Use ConcurrentHashMap in the VirtualNodeInfoProcessor (#2524)
If the VirtualNodeInfoProcessor is closed at the same time as it gets an update, a ConcurrentModificationException could occur as an iteration over a synchronizedMap occurs at the same time as trying to clear it. The map has been changed to a ConcurrentHashMap, which should guard against this exception in this case.