CORE-8114 Increase timeouts on subscription tests (#2544)
Two subscription test groups had timeouts an order of magnitude lower than the others. One of them is timing out. This only started happening after a timeout was introduced, if the test was failing previously it would have manifested itself as hanging. We never saw this, so most likely the new failure is only down to a very restrictive timeout. It's now aligned with the other tests of the same nature.