CORE-8042: Ensure RegistrationServiceLifecycleHandler closes resources in the right order (#2538)
RegistrationServiceLifecycleHandler is responsible for the lifecycle of the membership registration service. The lifecycle handler creates a compacted subscription and then listens to it in order to respond to its lifecycle changes. However, when the subscription is created, previously no attempt was made to close the registration handle for that subscription. On closing the old subscription, a RegistrationStatusChangeEvent would be posted to coordinator indicating that the old subscription had gone down, which then resulted in the component shutting down completely. This situation cannot be recovered, as the dependency is circular - the component is waiting for the subscription to come up, but that can only happen if the component goes up and recreates it.

This PR ensures the registration handle is closed before closing the subscription on config change. The handler has also been migrated to use the managed resources API, which ensures objects are closed if they are ever recreated. This in turn reduces clutter in the class as a whole. Note that this is not appropriate for the publisher as this is exposed outside the class. Additionally the tests have been migrated to use the lifecycle testing API, to assist in readability.