CORE-7261 Refactor group policy parsing when CPI metadata is unavailable (#2528)
When a holding identity modifies their virtual node information, the group policy for that holding identity is parsed again - in case the virtual node change affected the group policy file. For this, the `GroupPolicyProvider` fetches the CPI metadata from the `CpiInfoReadService`. If this metadata is not yet available, the `GroupPolicyProvider` previously would proceed to attempt processing the group policy, and log error messages since the group policy would inevitably be null. After this change, the `GroupPolicyProvider` falls back on an existing mechanism to process the group policy on the next read instead. This change also improves the logging around this scenario.