CORE-8018 Fix dropped registration ID from RPC worker (#2542)
When a registration request is submitted, the RPC worker generates an ID for the request. When the membership worker processes that request, it generates a new ID. In error scenarios, the RPC worker cannot return the registration ID since it has changed.
This change addresses the issue so that the ID is generated once in the RPC worker, and used by the membership worker.