CORE-7298 Virtual node vault database reconcile reader for group parameters tables (#2453)
This change adds the database reconcile reader used for reconciling group parameters database tables with kafka records. The group parameters kafka reconcile reader/writer, and the main reconciler class are out of scope for this commit.

This extracts the handling of the entity manager factory and entity manager creation from DbReconcilerReader so that implementations of ReconcilerWrapper can determine how/when to create/close the entity manager factory. The changes in this commit also allows implementations of the wrapper class to specify the lifecycle dependencies. After these changes, it is necessary to provide a ReconciliationContext factory to the DbReconcilerReader which creates a list of reconciliation contexts to be used during reconciliation. EMF and EM creation and closing are handled by implementations of ReconciliationContext. ClusterReconciliationContext covers existing reconciliation implementations which uses an existing EMF to create and close an EM per reconciliation, and VirtualNodeReconciliationContext covers creating and closing the EMF and EM per reconciliation.

Existing reader implementations have been updated to work with this refactoring. The main reason for the refactoring is the addition of the GroupParametersReconciler reader. This reader requires reading per vnode since it works against the virtual node vault databases. This class attempts to create an reconciliation context for each virtual node during reconciliation, create a stream of results for each vault DB, combine the results into a single stream, and when the stream of data is consumed the reconciliation context is closed. It returns a stream containing the most recent group parameters in the context of the virtual nodes which were reconciling.