CORE-6010 fix flakey test by verifying the outputs of a transaction rather than only waiting on a latch (#2587)
This test was flakey as previously a number of records were published and a latches count was checked before publishing more records. Given that 2 consumers in the same group are started within the same process there is a race condition in the repartitioning logic. One consumer may be given both partitions and poll before the second consumer starts, triggering repartition event. This means that an additional 5 counts of the latch were seen in some runs.

Change the test to have the processors send outputs to a different topic and verify all the records are sent there.