CORE-7089 Distribute group parameters after registration approval (#2473)
The MGM now distributes updated membership packages in response to a new `DistributeMembershipPackage` command - this was previously done during registration approval. The new command is issued after registration approval, and the membership packages distributed by the MGM now include the signed group parameters. The group parameters are also published to Kafka whenever the MGM persists the initial group parameters snapshot or publishes the membership package.
Changes in response to the altered `PersistGroupParametersResponse` schema are also included.