Lifecycle tweaks to prevent using closed publisher
Explicitly sets writer components as down during states of config
updates.

This should prevent attempted publishes while the publisher it requires
is in the process of being recreated with a new messaging config.

This change additionally ensures that closed resources are set to null,
and that the resources aren't cleaned up too aggressively in the case of
a config update