CORE-7299 - Integrate reconciliaton services for group parameters (#2611)
Created final version for group parameters reconciliation.
Also this PR fixes a bug found in PersistGroupParametersHandler. We used .singleResult which assumes that the db already has some kind of data stored. When we tried to persist the group parameters for member for the first time we ended up with: No entity found for query, because the db table was empty and no entry was found. I added an extra check to the places where we are using .singleResult to make sure the query from db did not resulted in an empty list before validations.