CORE-8417 Send the correct certificate during session negotation
When processing an initiator hello message the link manager looks
up the first locally hosted identity that is in the same group
as the sender of the initiator hello message. We then use this
identity as a part of the responder handshake message. However
this is not  necessarily the identity that the initiator intended
to communicate with. The intended identity is sent as a part of
the initiator handshake message. So use this instead to look up
the correct certificate to send to the initiator.