CORE-8413 Fix intermittently failing group parameters cache test (#2624)
`GroupParametersCacheTest` was failing intermittently on Windows builds, as it expected the modified time group parameter to be strictly before a time instant specified in the test - and did not take into account the two time instants being equal. This change fixes the above, and also applies the same update to `GroupParametersFactoryTest`.