CORE-8417 Send the correct certificate during session negotation (#2631)
Change the Link Manager to always use the correct certificate. When processing an initiator hello message the link manager looks up the first locally hosted identity that is in the same group as the sender of the initiator hello message. We were previously using this identity as a part of the responder handshake message. This is not necessarily the identity that the initiator intended to communicate with. The actual intended identity is sent as a part of the initiator handshake message. So we now use this identity instead to look up the correct certificate to send to the initiator.